list of changes: Rooms now have Ids. A room's Id is the # of servers made since the start of the server ie. the first server has Id #1, the second has Id #2
Upon initialization of the server engine, room 'main' gets created.
It has a dummy host of 'admin' and has a roomId of 0
The admin account is not available for use, nor is it counted as a user.

All instances of 'currentRoom=nullptr' have been changed to 'currentRoomId==0'

User messages are now set to use sendRoomMessage

ServerEngine::registerRoom has been changed to take in the room creators ID as an argument.

/info crash is now fixed.